requirements.txt
Django==4.0
django-extensions==3.1.5
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
pandas==1.3.5
coverage==6.2
PyVCF==0.4.3

